Limit guitar-part duplication with "tag" mechanism
When writing guitar parts it is useful to add commands that will
influence generated tablatures such as `\set TabStaff.minimumFret` or
string number indications.

However, if one wants to typeset a score without a tablature staff,
these commands add either ghost tablatures or string numbering above the
score which are not desirable (certainly the ghost tablatures).

Prior to this change, the solution was to use duplicated guitar parts,
one of which had tabalature indications. Any musical change in one had
to be repeated in the other.

We can use Lilypond's `\tag` mechanism to wrap the commands and then use
the `\removeWithTag` to filter them in the `\score` section.

For the `\set TabStaff` commands a single tag is sufficient as they are
not directly linked to a note.

For string numbering indications, the notes or whole phrase must be
tagged and a non-numbered version must be added and tagged. This does represent
some duplication but hopefully much more limited, and any musical
modification is easier because the expressions will be next to each
other.

We use the `\tag #'tabs` for tablature indications and `\tag #'plain`
for the alternative.

* remove all duplicate guitar parts and tag tablature indications
* adapt score output commands to filter tags appropriately